HealthAI/requirements.txt
streamlit>=1.31.1
openai>=1.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2